@@ -1,11 +1,9 @@
 name: progress_dialog
 description: A light weight library to easily manage a progress dialog with simple steps whenever you need to do it. You can easily show and hide it.
-version: 1.2.4
-#author: Fayaz<fayazfz07@gmail.com>
-homepage: https://github.com/fayaz07/progress_dialog
+version: 2.0.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.7.0"
 
 dependencies:
   flutter:
