@@ -1,13 +1,11 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart';
 
 enum ProgressDialogType { Normal, Download }
 
 String _dialogMessage = "Loading...";
 double _progress = 0.0, _maxProgress = 100.0;
 
-Widget _customBody;
+Widget? _customBody;
 
 TextAlign _textAlign = TextAlign.left;
 Alignment _progressWidgetAlignment = Alignment.centerLeft;
@@ -15,14 +13,14 @@ Alignment _progressWidgetAlignment = Alignment.centerLeft;
 TextDirection _direction = TextDirection.ltr;
 
 bool _isShowing = false;
-BuildContext _context, _dismissingContext;
-ProgressDialogType _progressDialogType;
+late BuildContext _context, _dismissingContext;
+ProgressDialogType _progressDialogType = ProgressDialogType.Normal;
 bool _barrierDismissible = true, _showLogs = false;
 
 TextStyle _progressTextStyle = TextStyle(
-        color: Colors.black, fontSize: 12.0, fontWeight: FontWeight.w400),
-    _messageStyle = TextStyle(
-        color: Colors.black, fontSize: 18.0, fontWeight: FontWeight.w600);
+    color: Colors.black, fontSize: 12.0, fontWeight: FontWeight.w400);
+TextStyle _messageStyle = TextStyle(
+    color: Colors.black, fontSize: 18.0, fontWeight: FontWeight.w600);
 
 double _dialogElevation = 8.0, _borderRadius = 8.0;
 Color _backgroundColor = Colors.white;
@@ -35,37 +33,37 @@ Widget _progressWidget = Image.asset(
 );
 
 class ProgressDialog {
-  _Body _dialog;
+  late _Body _dialog;
 
   ProgressDialog(BuildContext context,
-      {ProgressDialogType type,
-        bool isDismissible,
-        bool showLogs,
-        TextDirection textDirection,
-        Widget customBody}) {
+      {ProgressDialogType? type,
+        bool? isDismissible,
+        bool? showLogs,
+        TextDirection? textDirection,
+        Widget? customBody}) {
     _context = context;
     _progressDialogType = type ?? ProgressDialogType.Normal;
     _barrierDismissible = isDismissible ?? true;
     _showLogs = showLogs ?? false;
-    _customBody = customBody ?? null;
+    _customBody = customBody;
     _direction = textDirection ?? TextDirection.ltr;
   }
 
   void style(
-      {Widget child,
-      double progress,
-      double maxProgress,
-      String message,
-      Widget progressWidget,
-      Color backgroundColor,
-      TextStyle progressTextStyle,
-      TextStyle messageTextStyle,
-      double elevation,
-      TextAlign textAlign,
-      double borderRadius,
-      Curve insetAnimCurve,
-      EdgeInsets padding,
-      Alignment progressWidgetAlignment}) {
+      {Widget? child,
+        double? progress,
+        double? maxProgress,
+        String? message,
+        Widget? progressWidget,
+        Color? backgroundColor,
+        TextStyle? progressTextStyle,
+        TextStyle? messageTextStyle,
+        double? elevation,
+        TextAlign? textAlign,
+        double? borderRadius,
+        Curve? insetAnimCurve,
+        EdgeInsets? padding,
+        Alignment? progressWidgetAlignment}) {
     if (_isShowing) return;
     if (_progressDialogType == ProgressDialogType.Download) {
       _progress = progress ?? _progress;
@@ -88,12 +86,12 @@ class ProgressDialog {
   }
 
   void update(
-      {double progress,
-      double maxProgress,
-      String message,
-      Widget progressWidget,
-      TextStyle progressTextStyle,
-      TextStyle messageTextStyle}) {
+      {double? progress,
+        double? maxProgress,
+        String? message,
+        Widget? progressWidget,
+        TextStyle? progressTextStyle,
+        TextStyle? messageTextStyle}) {
     if (_progressDialogType == ProgressDialogType.Download) {
       _progress = progress ?? _progress;
     }
@@ -117,22 +115,22 @@ class ProgressDialog {
         _isShowing = false;
         Navigator.of(_dismissingContext).pop();
         if (_showLogs) debugPrint('ProgressDialog dismissed');
-        return Future.value(true);
+        return true;
       } else {
         if (_showLogs) debugPrint('ProgressDialog already dismissed');
-        return Future.value(false);
+        return false;
       }
     } catch (err) {
       debugPrint('Seems there is an issue hiding dialog');
       debugPrint(err.toString());
-      return Future.value(false);
+      return false;
     }
   }
 
   Future<bool> show() async {
     try {
       if (!_isShowing) {
-        _dialog = new _Body();
+        _dialog = _Body();
         showDialog<dynamic>(
           context: _context,
           barrierDismissible: _barrierDismissible,
@@ -141,20 +139,21 @@ class ProgressDialog {
             return WillPopScope(
               onWillPop: () async => _barrierDismissible,
               child: Dialog(
-                  backgroundColor: _backgroundColor,
-                  insetAnimationCurve: _insetAnimCurve,
-                  insetAnimationDuration: Duration(milliseconds: 100),
-                  elevation: _dialogElevation,
-                  shape: RoundedRectangleBorder(
-                      borderRadius:
-                          BorderRadius.all(Radius.circular(_borderRadius))),
-                  child: _dialog),
+                backgroundColor: _backgroundColor,
+                insetAnimationCurve: _insetAnimCurve,
+                insetAnimationDuration: const Duration(milliseconds: 100),
+                elevation: _dialogElevation,
+                shape: RoundedRectangleBorder(
+                  borderRadius: BorderRadius.all(Radius.circular(_borderRadius)),
+                ),
+                child: _dialog,
+              ),
             );
           },
         );
         // Delaying the function for 200 milliseconds
         // [Default transitionDuration of DialogRoute]
-        await Future.delayed(Duration(milliseconds: 200));
+        await Future.delayed(const Duration(milliseconds: 200));
         if (_showLogs) debugPrint('ProgressDialog shown');
         _isShowing = true;
         return true;
@@ -225,11 +224,12 @@ class _BodyState extends State<_Body> {
             Row(
               children: <Widget>[
                 Expanded(
-                    child: Text(
-                      _dialogMessage,
-                      style: _messageStyle,
-                      textDirection: _direction,
-                    )),
+                  child: Text(
+                    _dialogMessage,
+                    style: _messageStyle,
+                    textDirection: _direction,
+                  ),
+                ),
               ],
             ),
             SizedBox(height: 4.0),
